Convert Luthas plugin from kts to kt so that companion objects can be used to reference player attributes
